pyproject.toml
[tool.poetry]
name = "order-book"
version = "0.1.0"
description = "task azati - Python Back-end Developer Assessment Task - Anžhalika Hrytsevich"
authors = ["Aliaksei Saukin <88047550+SaxleS@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
uvicorn = {extras = ["standard"], version = "^0.31.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
